% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_demo_data.R
\name{demo_data}
\alias{demo_data}
\title{This function produces survival demo data}
\usage{
demo_data(hazard = 1.2, censor_hazard = 1, pts = 160, maxtime = 4, arm = 1)
}
\arguments{
\item{hazard}{is the survival hazard}

\item{censor_hazard}{is the hazard used to censor data}

\item{pts}{is the number of patient in the data}

\item{maxtime}{is the tiem after which all events are censored}

\item{arm}{is a label add to all the data}
}
\value{
return a data frame with the survival demo data
}
\description{
This function produces survival demo data
}
\examples{
demo_data(hazard=1.5,censor_hazard=.9,pts=150,maxtime=3,arm=2)

}
