\name{example_data}
\alias{example_data}
\docType{data}
\title{
Example data for calculating mean and variance
}
\description{
This is an example data.
}
\usage{data("example_data")}
\format{
  A umeric list of several observations.
}
\details{
A set of observation of computing meam and variance
}
\source{
An example dataset.
}
\references{
R package TestRpackage
}
\examples{
data(example_data)
## maybe str(example_data) ; plot(example_data) ...
}
\keyword{datasets}
