% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft.R
\name{hard}
\alias{hard}
\title{Hard-thresholding function}
\usage{
hard(a, lambda, y = c(1, 2, 3))
}
\arguments{
\item{a}{scalar input}

\item{lambda}{scalar thresholding parameter, nonnegative}

\item{y}{a random vector}
}
\description{
Hard-thresholding function
}
