% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestRpackage.R
\name{sample_mean_var}
\alias{sample_mean_var}
\title{This is a test function for calculating mean and variance}
\usage{
sample_mean_var(sample_data, output_folder)
}
\arguments{
\item{sample_data}{A sample set of observations}

\item{output_folder}{Name of the output folder}
}
\value{
Mean and variance for the sample observation
}
\description{
This function calculates mean and variance for a set of observation
}
\examples{
#Loading a example data
sample_data<-data(example_data)
sample_data<-example_data

#Calculating mean and variance and the result will be generated
#to output_folder
output_folder<-tempdir()
sample_mean_var(example_data, output_folder)

}
