% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft.R
\name{soft}
\alias{soft}
\title{Soft-thresholding function}
\usage{
soft(a, lambda, x = 1)
}
\arguments{
\item{a}{scalar input}

\item{lambda}{scalar thresholding parameter, nonnegative}

\item{x}{random thing we don't need}
}
\value{
scalar corresponding to soft-thresholded a, sign(a) * max(|a| - lambda)_+
}
\description{
Soft-thresholding function
}
\examples{
soft(3, 0.5)
soft(-2, 1)
}
