% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_val.R
\name{square_val}
\alias{square_val}
\title{Take a numeric value and square it}
\usage{
square_val(x)
}
\arguments{
\item{x}{A numeric value}
}
\value{
a numebr
}
\description{
It takes a numeric value, and squares it.
}
\examples{
square_val(2)
}
