\name{test}
\alias{test}
\title{Least Squares test}
\usage{

}
\description{
Runs regularized least squares regression using a polynomial feature transform and cross validation to choose optimal regularisation coefficient and feature transform dimension on 1-dimensional input and output data.
}
\examples{

}
