% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function}
\alias{test_function}
\title{A test function}
\usage{
test_function(x = 3, y = 4)
}
\arguments{
\item{x}{One number}

\item{y}{another number}
}
\value{
x multipled by y
}
\description{
A description for this function
}
\details{
Plots the posterior change points

Add description
}
\examples{
test_function(x = 5, y = 7)
}
