\name{TestRPackage-package}
\alias{TestRPackage-package}
\alias{TestRPackage}
\docType{package}
\title{
\packageTitle{TestRPackage}
}
\description{
\packageDescription{TestRPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TestRPackage}
\packageIndices{TestRPackage}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{TestRPackage}

Maintainer: \packageMaintainer{TestRPackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
